Initial split types and namespaces refactoring
