+++ Modules/AnimeStatus.cs
@@ -1,0 +1,11 @@
+namespace YO.Modules
+{
+	/// <summary>
+	/// Anime release status.
+	/// </summary>
+	public enum AnimeStatus {
+		Released,
+		RegularOngoing,
+		PendingOngoing
+	};
+}
