+++ Modules/EntryMode.cs
@@ -1,0 +1,12 @@
+namespace YO.Modules
+{
+	/// <summary>
+	/// Entry action mode.
+	/// </summary>
+	public enum EntryMode {
+		Default,
+		AddNew,
+		AlreadyWatched,
+		Remove
+	};
+}
