+++ Windows/AboutWindow.xaml.cs
@@ -4,7 +4,7 @@ using System.Windows.Controls;
 using System.Windows.Media;
 using System.Windows.Media.Animation;
 
-namespace YO {
+namespace YO.Windows {
     /// <summary>
     /// Interaction logic for AboutWindow.xaml
     /// </summary>
