+++ Windows/MainWindow.xaml
@@ -1,10 +1,11 @@
-﻿<Window x:Name="mainWindow" x:Class="YO.MainWindow"
+﻿<Window x:Name="mainWindow" x:Class="YO.Windows.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:scm="clr-namespace:System.ComponentModel;assembly=WindowsBase"
         xmlns:local="clr-namespace:YO"
+        xmlns:windows="clr-namespace:YO.Windows"
         mc:Ignorable="d"
         Title="YO: Твои Онгоинги" Height="600" Width="1070" MinHeight="235" MinWidth="520" FontFamily="/YO;component/Resources/fonts/#Exo 2" FontSize="16" WindowState="Normal" Closing="Window_Closing" Icon="../Resources/img/ico/normal_ni.ico" Closed="Window_Closed" Loaded="MainWindow_Loaded" ShowInTaskbar="True">
     <Window.Resources>
@@ -17,11 +18,11 @@
         <Style TargetType="ScrollBar"/>
     </Window.Resources>
     <Window.CommandBindings>
-        <CommandBinding Command="{x:Static local:Commands.F5}" Executed="RefreshBarItem_Click"/>
-        <CommandBinding Command="{x:Static local:Commands.Esc}" Executed="SignInButtonEsc_Click"/>
-        <CommandBinding Command="{x:Static local:Commands.CtrlO}" Executed="ShikiBarItem_Click"/>
-        <CommandBinding Command="{x:Static local:Commands.CtrlM}" Executed="ViewBarItem_Click"/>
-        <CommandBinding Command="{x:Static local:Commands.CtrlQ}" Executed="ExitBarItem_Click"/>
+        <CommandBinding Command="{x:Static windows:Commands.F5}" Executed="RefreshBarItem_Click"/>
+        <CommandBinding Command="{x:Static windows:Commands.Esc}" Executed="SignInButtonEsc_Click"/>
+        <CommandBinding Command="{x:Static windows:Commands.CtrlO}" Executed="ShikiBarItem_Click"/>
+        <CommandBinding Command="{x:Static windows:Commands.CtrlM}" Executed="ViewBarItem_Click"/>
+        <CommandBinding Command="{x:Static windows:Commands.CtrlQ}" Executed="ExitBarItem_Click"/>
     </Window.CommandBindings>
     <Grid>
         <Grid x:Name="mainGrid">
