+++ Windows/Mode.cs
@@ -1,0 +1,12 @@
+namespace YO.Windows
+{
+	/// <summary>
+	/// Basic app states.
+	/// </summary>
+	enum Mode {
+		Normal,
+		SignIn,
+		Refresh,
+		Offline
+	};
+}
