+++ Windows/SortingMode.cs
@@ -1,0 +1,12 @@
+namespace YO.Windows
+{
+	/// <summary>
+	/// Sorting of animes in calendar slots and list.
+	/// </summary>
+	enum SortingMode {
+		Score,
+		Year,
+		Progress,
+		Alphabet
+	}
+}
