+++ Windows/StringResources.cs
@@ -1,0 +1,12 @@
+namespace YO.Windows
+{
+	/// <summary>
+	/// Massive string resources class.
+	/// </summary>
+	public static class StringResources {
+		public static string Help => "С чего начать?\n\nРасписание просмотра серий аниме составляется автоматически, основываясь на Вашем списке \"Смотрю\". Мы принципиально не просим пароль от Вашего аккаунта, поэтому поддерживаются только открытые списки. Изменить видимость списка можно в настройках Шикимори.\n\nИсключить тайтл из расписания и изменить период выхода серий (по умолчанию раз в неделю) можно, нажав на карандашик в правом верхнем углу обложки. Чтобы вернуть исключенное аниме, перейдите в Режим списка (Ctrl+M), там же доступен удобный интерфейс для управления параметрами всех Ваших аниме одновременно. Зеленым цветом подсвечиваются тайтлы, план по которым был выполнен сегодня, желтым — готовые к просмотру сегодня, красным цветом — просроченные. \n\nОдна из ключевых настроек программы — задержка онгоингов. Хотя в Японии новые серии можно посмотреть по телевизору в день выхода, для зарубежных зрителей это не так просто, поэтому по умолчанию установлена однодневная задержка между выходом новой серии и ее положением в расписании. Изменить период выхода и положение в расписании для текущих онгоингов с известными датами выхода серий можно, нажав кнопку \"Догнать\" в окошке редактирования.\n\nВы можете отключить автозапуск программы в настройках, но тогда Вы не будете получать уведомления, пока не запустите ее самостоятельно.\n\n\nЧАВО:\n\nВопрос: После некоторого времени использования программы расписание стало несбалансированным. Что делать?\nОтвет: Для тайтлов, которые стоят в неудобные дни, можно нажать кнопку \"Сброс\" в окне редактирования. Если все совсем плохо, воспользуйтесь \"Действия → Перестроить расписание\". Расписание будет составлено заново, но при этом вы потеряете все данные о текущем графике, включая задолженности!\n\nВопрос: Я посмотрел больше серий, чем нужно по плану, почему приложение не показывает, что я опережаю график?\nОтвет: Планировщик учитывает это, и обновляет прогресс тайтла так, чтобы Вы досмотрели его быстрее. :)\n\nВопрос: Приложение не запускается.\nОтвет: Проверьте значок программы в блоке фоновых приложений на панели задач. Для запуска кликните по нему два раза.\n\nВопрос: Вместо обложки аниме отображается 404-тян, хотя у тайтла есть постер на Шикимори. Как это исправить?\nОтвет: Скорее всего, это произошло из-за обрыва связи при синхронизации. Выйдите из программы (Ctrl+Q) , скопируйте в адресную строку проводника \"%APPDATA%/YO\", оказавшись там, удалите папку \"tmp\". Остальные файлы трогать не стоит, если не хотите потерять свои данные. ;)\n\nВопрос: ЧАВО не помогает, произошла неизвестная ошибка, программа упала, я нашел баг, опечатку.\nОтвет: Это замечательно! Пожалуйста, напишите о Вашей проблеме в топик ошибок, и она будет исправлена.\n\nСообщить об ошибке или предложить идею по улучшению можно в клубе темы (ссылку можно выделить и скопировать):\nhttps://shikimori.one/clubs/3417\n\nРепозиторий с исходным кодом, если Вы хотите поменять программу по своему усмотрению:\nhttps://github.com/SergejVolkov/YO";
+		public static string About => "YO: Твои Онгоинги — это удобный и функциональный планировщик просмотра аниме, синхронизирующийся с Вашим списком на Шикимори. С YO Вы не только можете смотреть релизнутые тайтлы так, как будто они выходят сейчас, но и планировать просмотр текущих онгоингов, догонять их и многое другое!\n\nВыбирайте периодичность выхода серий и их количество, а умный алгоритм составит расписание наиболее эффективным образом. Каждый день приложение будет рассылать уведомления с указанием эпизодов на сегодня. Вы также можете настроить для тайтлов любые ссылки на онлайн-кинотеатры: запланированные серии всего в одном клике от Вас! А если пропустили несколько эпизодов — ничего страшного: просроченные тайтлы можно обнулить или исключить из расписания.\n\nВ приложении доступна темная тема, переключение между различными аккаунтами. Мы не собираем о Вас никакие данные, все хранится только на Вашем компьютере.";
+		public static string Copyright => "YO: Your Ongoings is a strictly non-profit open-source project distributed under GNU GPLv3 (GNU General Public License 3.0). The author does not own any of the drawings and images that are used in this program except the logo icon, however, neither this program nor any of its derivatives shall not gain profit of them, so the inclusion of the aforementioned images can be considered fair use.\n\nIf you hold the rights to any of the images in this program and would like to stop their use, please contact the author.\n\nE-mail: volkovsergej17@gmail.com\nTelegram: @SergejVolkov";
+		public static string Notice => "SergejVolkov, 2021";
+	}
+}
