+++ Windows/TitleLanguage.cs
@@ -1,0 +1,10 @@
+namespace YO.Windows
+{
+	/// <summary>
+	/// Language of anime titles.
+	/// </summary>
+	public enum TitleLanguage {
+		Russian,
+		Romaji
+	};
+}
